+++ app/main/event.js
@@ -2,39 +2,12 @@
  * @description 主进程事件监听
  */
 
-import { ipcMain, BrowserWindow, app, Menu, dialog } from 'electron';
+import { ipcMain, BrowserWindow, app, Menu, dialog, shell } from 'electron';
 import fs from 'fs';
 import fse from 'fs-extra';
-import marked from 'marked';
-import html2pdf from 'html-pdf';
+import { markedToHtml } from '../views/utils/utils';
 import schedule from './schedule';
-import pdfAddStyle from './pdf';
-
-const renderer = new marked.Renderer();
-
-renderer.listitem = function (text) {
-  let res = text;
-  if (/^\s*\[[x ]\]\s*/.test(text)) {
-    res = text.replace(/^\s*\[ \]\s*/, '<input class="task-list-item-checkbox" type="checkbox" disabled></input> ').replace(/^\s*\[x\]\s*/, '<input class="task-list-item-checkbox" checked disabled type="checkbox"></input> ');
-    return `<li class="task-list-li">${res}</li>`;
-  }
-  return `<li>${text}</li>`;
-};
-
-marked.setOptions({
-  renderer,
-  gfm: true,
-  tables: true,
-  breaks: false,
-  pedantic: false,
-  sanitize: false,
-  smartLists: true,
-  smartypants: false,
-  highlight: (code) => {
-    const value = require('../views/utils/highlight.min.js').highlightAuto(code).value;
-    return value;
-  },
-});
+import PDF from './PDF';
 
 const dataPath = app.getPath('appData');
 let appDataPath = `${dataPath}/Yosoro`;
@@ -442,6 +415,7 @@ export function eventListener(menus) {
 
   ipcMain.on('export-note', (event, args) => {
     const { projectName, fileName, type, data } = args;
+    const folderPath = `${projectsPath}/${projectName}`;
     const filePath = `${projectsPath}/${projectName}/${fileName}.md`;
     try {
       let content;
@@ -458,42 +432,28 @@ export function eventListener(menus) {
       } else if (type === 'html') {
         title = 'Export as Html';
         if (!data) {
-          content = marked(content);
+          content = markedToHtml(content, false);
         }
       } else if (type === 'pdf') {
         title = 'Export as PDF';
-        if (!data) {
-          content = marked(content);
-        }
       }
       const options = {
         title,
         defaultPath: `${app.getPath('desktop')}/${fileName}.${type}`,
       };
-      dialog.showSaveDialog(options, (filename) => {
-        if (typeof filename === 'string') {
+      dialog.showSaveDialog(options, async (fname) => {
+        if (typeof fname === 'string') {
           const extension = `.${type}$`;
           const reg = new RegExp(extension, 'ig');
-          let file = filename;
-          if (!reg.test(filename)) {
+          let file = fname;
+          if (!reg.test(fname)) {
             file += `.${type}`;
           }
-          if (type === 'pdf' && content) {
+          if (type === 'pdf') {
             event.sender.send('async-export-file');
-            content = pdfAddStyle(content);
-            const windowToPDF = new BrowserWindow({ show: false });
-            const tempPath = `${app.getPath('temp')}/yosoro_pdf.html`;
-            fs.writeFileSync(tempPath, content);
-            windowToPDF.loadURL(`file://${tempPath}`);
-            setTimeout(() => {
-              windowToPDF.webContents.printToPDF({}, (err, pdfData) => {
-                if (err) {
-                  throw err;
-                }
-                fs.writeFileSync(file, pdfData);
-                event.sender.send('async-export-file-complete');
-              });
-            }, 1000);
+            const pdf = new PDF([`${fileName}.md`], folderPath, file, false);
+            await pdf.start();
+            event.sender.send('async-export-file-complete');
           } else {
             fs.writeFileSync(file, content);
           }
@@ -519,6 +479,10 @@ export function eventListener(menus) {
       }
       if (type === 'md') {
         fse.copySync(folderPath, exportPath);
+      } else if (type === 'pdf') {
+        const notes = fs.readdirSync(folderPath);
+        const pdf = new PDF(notes, folderPath, exportPath);
+        await pdf.start();
       } else {
         const promiseArr = [];
         const notes = fs.readdirSync(folderPath);
@@ -526,32 +490,21 @@ export function eventListener(menus) {
           let content = fs.readFileSync(`${folderPath}/${note}`, {
             encoding: 'utf8',
           });
-          content = marked(content);
+          content = markedToHtml(content, false);
           const name = note.replace(/\.md/ig, '');
-          if (type === 'pdf' && content) {
-            promiseArr.push(new Promise((resolve) => {
-              content = pdfAddStyle(content);
-              html2pdf.create(content).toFile(`${exportPath}/${name}.${type}`, (err) => {
-                if (err) {
-                  console.warn(err);
-                }
-                resolve('done');
-              });
-            }));
-          } else if (type === 'html') {
-            promiseArr.push(new Promise((resolve) => {
-              fs.writeFile(`${exportPath}/${name}.${type}`, content, (err) => {
-                if (err) {
-                  console.warn(err);
-                }
-                resolve('done');
-              });
-            }));
-          }
+          promiseArr.push(new Promise((resolve) => {
+            fs.writeFile(`${exportPath}/${name}.${type}`, content, (err) => {
+              if (err) {
+                console.warn(err);
+              }
+              resolve('done');
+            });
+          }));
         }
         await Promise.all(promiseArr);
       }
       event.sender.send('async-export-file-complete');
+      shell.openExternal(`file://${exportPath}`);
     } catch (ex) {
       console.warn(ex);
       event.sender.send('async-export-file-complete');